Julia: Move backup line to package and disable sed command
First to avoid warning from makepkg about missing backup file in package
and the second to avoid a faulty call to sphinx-build2. By default the html will be build, and julia downloads all the stuff it needs (sphinx, the theme) in a virtualenv. With the sed command active the build process will explode.

Forgot to remove one versionstring