Sylpheed-light: Add validpgpkeys
The released tarballs are signed. With the pacman4.2 release the behaviour for the .asc and co has changed.

Quote: "Source signatures are required to be from a trusted source or listed in the validpgpkeys array."

So I added the key of Hiroyuki Yamamoto as he signs the releases with this key