Patches: Add patches for custmized AUR and repo packages
Previously I just added customized repo packages to this repo (because of ABS) and ignored changes I made to AUR packages.
Then I added both (or not).
Switched to ASP and diffed still by hand instead of adding it directly to the asp repos. Then I did this for some packages and didn't track them in this repo (mpv, mpd).

So now I let the changes where they belong and just track here the patches and the PKGBUILD I maintain on the AUR.
Exceptions are made for linux and co.