tmsu: make it compile again
Newer Go versions deprecated the old way of obtaining modules. Since then projects need the Go mod and sum files.
Normally there is a workaround to create those files on the fly, but it isn't applicable as tmsu with the latest release has an incompatible project structure.
To make this project compile again backport three upstream commits as patches.

Additionally update the fuse dep to fuse2 (was renamed in the past because of fuse3) and remove git as makedep.
Furthermore remove the provides line. In this case not needed.