opencloud-git: adjust make call
Build needs to be run in a specific subfolder.
Add -j1 as the generate part doesn't seem to be threadsafe. My makepkg.conf sets -j4 and then the build fails at varius and differing stages (most of the time due to binary not available which is currently in the process of being (re)installed)