opencloud: fix build
make build needs to be run in the subfolder opencloud.
Adjust therefore the command after a recommendation of upstream.
Additionally add -j1 as make generate doesn't seem to be threadsafe. I have -j4 set in makepkg.conf and the build will fail because of missing binaries that are in the process of being (re)installed.